Separate queued thread input actions
