+++ backend/threads/chat_adapters/thread_input_inlet.py
@@ -2,14 +2,16 @@ from __future__ import annotations
 
 from typing import Any
 
+from backend.threads.chat_adapters.queued_thread_input_action import (
+    dispatch_queued_thread_input_action,
+    queued_command_thread_input_action,
+    queued_thread_input_action,
+)
 from backend.threads.chat_adapters.runtime_action import dispatch_runtime_action
 from backend.threads.chat_adapters.runtime_thread_input_action import (
     RuntimeThreadInputAction,
-    dispatch_queued_thread_input_action,
     internal_runtime_thread_input_action,
     owner_runtime_thread_input_action,
-    queued_command_thread_input_action,
-    queued_thread_input_action,
 )
 from protocols.agent_runtime import AgentThreadInputResult
