+++ tests/Unit/integration_contracts/test_threads_router.py
@@ -13,15 +13,17 @@ import pytest
 from fastapi import Request
 from langchain_core.messages import HumanMessage, SystemMessage, ToolMessage
 
-from backend.threads.chat_adapters.runtime_thread_input_action import (
+from backend.threads.chat_adapters.queued_thread_input_action import (
     QueuedThreadInputAction,
-    RuntimeThreadInputAction,
     dispatch_queued_thread_input_action,
+    queued_command_thread_input_action,
+    queued_thread_input_action,
+)
+from backend.threads.chat_adapters.runtime_thread_input_action import (
+    RuntimeThreadInputAction,
     internal_runtime_thread_input_action,
     owner_runtime_thread_input_action,
     plan_runtime_thread_input_envelope,
-    queued_command_thread_input_action,
-    queued_thread_input_action,
 )
 from backend.threads.chat_adapters.thread_input_inlet import requeue_thread_input_item
 from backend.web.models.requests import CreateThreadRequest, ResolvePermissionRequest, SendMessageRequest, ThreadPermissionRuleRequest
